+++ src/components/deploymentConfig/types.tsx
@@ -71,3 +71,30 @@ export interface DeploymentConfigProps extends RouteComponentProps<DeploymentCon
     respondOnSuccess?: () => void;
 }
 
+export interface EsoData {
+    secretKey: string
+    key: string
+    property: string
+}
+export interface SecretFormProps {
+    id: number
+    appChartRef: { id: number; version: string; name: string }
+    appId: number
+    roleARN: string
+    name: string
+    index: number
+    external: boolean
+    externalType: string
+    secretData: { key: string; name: string; property: string; isBinary: boolean }[]
+    esoSecretData?: { secretStore: any; esoData: EsoData[] }
+    type: string
+    data: { k: string; v: string }[]
+    isUpdate: boolean
+    mountPath: string
+    keyValueEditable?: boolean
+    filePermission: string
+    subPath: boolean
+    update: (...args) => void
+    collapse: (...args) => void
+    initialise?: () => void
+}
